+++ sql/slugs.sql
@@ -10,6 +10,6 @@ CREATE TABLE IF NOT EXISTS slugs(
 	UNIQUE INDEX full_slug_idx(full_slug),
 	primary key(id)
 )
-ENGINE=InnoDB
+ENGINE=InnoDB ROW_FORMAT=DYNAMIC
 DEFAULT CHARSET=utf8mb4
 COLLATE=utf8mb4_general_ci;